added test in StorageDataBaseTest
