+++ src/main/resources/config.properties
@@ -9,4 +9,4 @@ storage_type=STORAGE_DATABASE
 db.connection_type=by_hibernate
 db.url=jdbc:postgresql://localhost:5432/storage_database
 db.username=postgres
-db.password=postgres
+db.password=riviera0685
